Add tool leprechaun to gnome farm patch

+++ src/script/packetdecoder/command/EmptyCommand.java
@@ -11,30 +11,30 @@ import lombok.var;
 public class EmptyCommand implements Command {
 
     @Override
-    public boolean canUse(Player player) {
-        return !player.getController().inWilderness() && !player.getController().inPvPWorld();
-    }
-
-    @Override
     public String getExample() {
         return "- Empties your inventory.";
     }
 
     @Override
     public void execute(Player player, String message) {
-        DialogueEntry entry = new DialogueEntry();
-        entry.setSelection("Are you sure you want to empty your inventory?", "Yes, empty my inventory!", "No!");
-        DialogueScript script = (p, index, childId, slot) -> {
-            for (var i = 0; i < player.getInventory().size(); i++) {
-                var id = player.getInventory().getId(i);
-                if (slot == 0) {
-                    player.getInventory().deleteItem(id, Item.MAX_AMOUNT);
-                } else {
-                    return;
+        if (!player.getController().inWilderness() && !player.getController().inPvPWorld()) {
+            DialogueEntry entry = new DialogueEntry();
+            entry.setSelection("Are you sure you want to empty your inventory?", "Yes, empty my inventory!", "No!");
+            DialogueScript script = (p, index, childId, slot) -> {
+                for (var i = 0; i < player.getInventory().size(); i++) {
+                    var id = player.getInventory().getId(i);
+                    if (slot == 0) {
+                        player.getInventory().deleteItem(id, Item.MAX_AMOUNT);
+                    } else {
+                        return;
+                    }
                 }
-            }
-            player.getGameEncoder().sendMessage("You empty your inventory..");
-        };
-        DialogueOld.open(player, entry, script);
+                player.getGameEncoder().sendMessage("You empty your inventory..");
+            };
+            DialogueOld.open(player, entry, script);
+        } else {
+            player.getGameEncoder().sendMessage("You can not use this command in the wilderness.");
+            return;
+        }
     }
 }
